Update makerdao_operations_may2021to2022_creationquery
Data stored on the blockchain is not human-readable and must be translated. Using the consistently updated public dataset for the ethereum blockchain from GoogleBigQuery, I parsed and annotated emitted event data for MakerDAO specific events. Not all events are annotated however, there was an effort to include most of the important ones. Annotations were sourced from etherscan and the MakerDAO technical docs. Note that values are stored in hexadecimal format as they are too large for BigQuery to store in numeric form.